fix(SonarCloud): resolve 17 issues — bugs, complexity, constants, StringBuilder
- S3923: conditional returns same value (already fixed by user)
- S3776: extract BuildTopMessage/AppendLocalTop/GetMedal to reduce cognitive complexity
- S1172: remove unused baseQuestion param from DecorateTokenQuestion
- S3358: extract GetActionWord to remove nested ternaries
- S1854: remove useless myPosition assignment
- S1905: remove unnecessary int cast in QuantitySuggestionService
- S1192: constants for RefreshButtonText, PositiveIntErrorMessage, YouMarker
- S1643: StringBuilder for profile in HandleGetProfile
- CA1816: add GC.SuppressFinalize in test Dispose

+++ ArkWallet/Application/Services/SuggestionServices/QuantitySuggestionService.cs
@@ -41,7 +41,7 @@ namespace ArkWallet.Application.Services.SuggestionServices
             if (portfolioItem == null || portfolioItem.Quantity <= 0)
                 return [];
 
-            var totalQuantity = (int)portfolioItem.Quantity;
+            var totalQuantity = portfolioItem.Quantity;
 
             var suggestions = Percentages
                 .Select(p => (int)Math.Floor(totalQuantity * p))
